+++ nginx-sys/src/http.rs
@@ -1,4 +1,4 @@
-use core::mem::offset_of;
+use ::core::mem::offset_of;
 
 use crate::bindings::ngx_http_conf_ctx_t;
